Feat(options): add confirmOnEnter option
if this option is set, when the remodal is open and user presses the "Enter" key, the remodal is closed with reason "STATE_CHANGE_REASONS.CONFIRMATION".
This may lead to unexpected behaviour if the remodal contains an input that also expects to do something on "Enter" key (e.g. a <textarea>).